lpcsdr_discover_devices: return type tweak

+++ include/lpcsdr.h
@@ -134,7 +134,7 @@ int lpcsdr_set_log_callback(lpcsdr_context *ctx, lpcsdr_log_callback callback);
 /* device.c */
 int lpcsdr_open_device(lpc_device *device, lpcsdr_device_handle **device_handle);
 void lpcsdr_free_device_list(lpc_device **device_list);
-int lpcsdr_discover_devices(lpcsdr_context *ctx, lpc_device ***lpc_device_list, bool allow_rom_bootloader);
+ssize_t lpcsdr_discover_devices(lpcsdr_context *ctx, lpc_device ***lpc_device_list, bool allow_rom_bootloader);
 int lpcsdr_open_single_device(lpcsdr_context *ctx, lpcsdr_device_handle **device_handle);
 int lpcsdr_close_device(lpcsdr_device_handle *dev);
